Change Fake to simple impl
